refactor(cloudformation): extract 6 resource families from provisioner mod.rs
resource_provisioner/mod.rs was 9003 lines: a single 219-method file whose core
dispatchers (create_resource / update_resource / delete_resource / get_att) were
buried among cohesive per-service handler families. The crate already splits most
services into submodules (acm.rs, apigw.rs, ...); this extends that convention to
six families that were still inline, each moved into its own submodule as an
'impl ResourceProvisioner' block (methods promoted to pub(crate) so the core
dispatchers still call them):

  elbv2 (18), cloudfront (18), elasticache (14), organizations (10),
  cwlogs (8), appautoscaling (4)

Pure code movement, no behavior change: mod.rs drops to 5876 lines, the 219-fn
set is identical before/after, 267 lib tests pass, workspace clippy -D warnings
clean. Submodules pick up the shared imports/helpers via 'use super::*'.
