+++ crates/fakecloud-cloudformation/src/resource_provisioner/mod.rs
@@ -1072,22 +1072,27 @@ mod acm;
 mod acmpca;
 mod apigw;
 mod apigwv2;
+mod appautoscaling;
 mod athena;
 mod autoscaling;
 mod batch;
 mod cloudformation;
+mod cloudfront;
 mod cloudwatch;
 mod codeartifact;
 mod codecommit;
 mod cognito;
 mod config;
+mod cwlogs;
 mod dynamodb;
 mod ec2;
 mod ecr;
 mod ecs;
 mod efs;
 mod eks;
+mod elasticache;
 mod elasticbeanstalk;
+mod elbv2;
 mod eventbridge;
 mod firehose;
 mod glue;
@@ -1099,6 +1104,7 @@ mod kms;
 mod lambda;
 mod logs;
 mod mq;
+mod organizations;
 mod pipes;
 mod rds;
 mod route;
@@ -2425,53 +2431,6 @@ impl ResourceProvisioner {
 
     // --- CloudWatch Logs ---
 
-    fn create_log_group(&self, resource: &ResourceDefinition) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let log_group_name = props
-            .get("LogGroupName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id);
-
-        let retention_in_days = props
-            .get("RetentionInDays")
-            .and_then(|v| v.as_i64())
-            .map(|v| v as i32);
-
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        let arn = format!(
-            "arn:aws:logs:{}:{}:log-group:{}:*",
-            state.region, state.account_id, log_group_name
-        );
-
-        let log_group = fakecloud_logs::LogGroup {
-            name: log_group_name.to_string(),
-            arn: arn.clone(),
-            creation_time: Utc::now().timestamp_millis(),
-            retention_in_days,
-            kms_key_id: None,
-            stored_bytes: 0,
-            log_streams: std::collections::BTreeMap::new(),
-            tags: std::collections::BTreeMap::new(),
-            subscription_filters: Vec::new(),
-            data_protection_policy: None,
-            index_policies: Vec::new(),
-            transformer: None,
-            deletion_protection: false,
-            log_group_class: Some("STANDARD".to_string()),
-        };
-
-        state
-            .log_groups
-            .insert(log_group_name.to_string(), log_group);
-        Ok(ProvisionResult::new(arn.clone()).with("Arn", arn))
-    }
-
-    /// Look up an S3 object's bytes from the in-process S3 state. Used by
-    /// the Lambda function provisioner to hydrate `Code.S3Bucket` /
-    /// `Code.S3Key` references into real ZIP content. Returns an error
-    /// string when the bucket or key is missing so the CFN error
-    /// surfaces back to the caller.
     fn read_s3_object_bytes(&self, bucket: &str, key: &str) -> Result<Vec<u8>, String> {
         let mut accounts = self.s3_state.write();
         let state = accounts.get_or_create(&self.account_id);
@@ -2672,871 +2631,6 @@ impl ResourceProvisioner {
 
     // --- ELBv2 ---
 
-    fn create_elbv2_load_balancer(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let scheme = props
-            .get("Scheme")
-            .and_then(|v| v.as_str())
-            .unwrap_or("internet-facing")
-            .to_string();
-        let lb_type = props
-            .get("Type")
-            .and_then(|v| v.as_str())
-            .unwrap_or("application")
-            .to_string();
-        let ip_address_type = props
-            .get("IpAddressType")
-            .and_then(|v| v.as_str())
-            .unwrap_or("ipv4")
-            .to_string();
-        let security_groups: Vec<String> = props
-            .get("SecurityGroups")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|s| s.as_str().map(|s| s.to_string()))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let tags = parse_elb_tags(props.get("Tags"));
-
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let lb_id = Uuid::new_v4().simple().to_string();
-        let arn = format!(
-            "arn:aws:elasticloadbalancing:{}:{}:loadbalancer/{}/{}/{}",
-            self.region,
-            self.account_id,
-            if lb_type == "network" { "net" } else { "app" },
-            name,
-            &lb_id[..16]
-        );
-        let dns_name = format!(
-            "{}-{}.{}.elb.{}.amazonaws.com",
-            name,
-            &lb_id[..16],
-            self.region,
-            self.region
-        );
-
-        let mut availability_zones: Vec<fakecloud_elbv2::AvailabilityZone> = Vec::new();
-        if let Some(arr) = props.get("Subnets").and_then(|v| v.as_array()) {
-            for s in arr {
-                if let Some(subnet_id) = s.as_str() {
-                    availability_zones.push(fakecloud_elbv2::AvailabilityZone {
-                        zone_name: format!("{}a", self.region),
-                        subnet_id: subnet_id.to_string(),
-                        outpost_id: None,
-                        load_balancer_addresses: Vec::new(),
-                        source_nat_ipv6_prefixes: Vec::new(),
-                    });
-                }
-            }
-        }
-
-        state.load_balancers.insert(
-            arn.clone(),
-            LoadBalancer {
-                arn: arn.clone(),
-                name: name.clone(),
-                dns_name: dns_name.clone(),
-                canonical_hosted_zone_id: "Z2P70J7EXAMPLE".to_string(),
-                created_time: Utc::now(),
-                scheme,
-                vpc_id: String::new(),
-                state_code: "active".to_string(),
-                state_reason: None,
-                lb_type,
-                availability_zones,
-                security_groups,
-                ip_address_type,
-                customer_owned_ipv4_pool: None,
-                enforce_security_group_inbound_rules_on_private_link_traffic: None,
-                enable_prefix_for_ipv6_source_nat: None,
-                ipv4_ipam_pool_id: None,
-                tags,
-                attributes: BTreeMap::new(),
-                minimum_capacity_units: None,
-                bound_port: None,
-            },
-        );
-
-        Ok(ProvisionResult::new(arn.clone())
-            .with("LoadBalancerArn", arn)
-            .with(
-                "LoadBalancerFullName",
-                format!("app/{name}/{}", &lb_id[..16]),
-            )
-            .with("LoadBalancerName", name)
-            .with("DNSName", dns_name)
-            .with("CanonicalHostedZoneID", "Z2P70J7EXAMPLE"))
-    }
-
-    fn delete_elbv2_load_balancer(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.load_balancers.remove(physical_id);
-        // Cascade-delete listeners and rules attached to this LB.
-        let listeners: Vec<String> = state
-            .listeners
-            .iter()
-            .filter(|(_, l)| l.load_balancer_arn == physical_id)
-            .map(|(arn, _)| arn.clone())
-            .collect();
-        for arn in &listeners {
-            state.listeners.remove(arn);
-            let rules: Vec<String> = state
-                .rules
-                .iter()
-                .filter(|(_, r)| r.listener_arn == *arn)
-                .map(|(a, _)| a.clone())
-                .collect();
-            for r in rules {
-                state.rules.remove(&r);
-            }
-        }
-        for tg in state.target_groups.values_mut() {
-            tg.load_balancer_arns.retain(|a| a != physical_id);
-        }
-        Ok(())
-    }
-
-    fn create_elbv2_target_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let protocol = props
-            .get("Protocol")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let port = props.get("Port").and_then(|v| v.as_i64()).map(|n| n as i32);
-        let vpc_id = props
-            .get("VpcId")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let target_type = props
-            .get("TargetType")
-            .and_then(|v| v.as_str())
-            .unwrap_or("instance")
-            .to_string();
-        let ip_address_type = props
-            .get("IpAddressType")
-            .and_then(|v| v.as_str())
-            .unwrap_or("ipv4")
-            .to_string();
-        let protocol_version = props
-            .get("ProtocolVersion")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let tags = parse_elb_tags(props.get("Tags"));
-
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let id = Uuid::new_v4().simple().to_string();
-        let arn = format!(
-            "arn:aws:elasticloadbalancing:{}:{}:targetgroup/{}/{}",
-            self.region,
-            self.account_id,
-            name,
-            &id[..16]
-        );
-
-        state.target_groups.insert(
-            arn.clone(),
-            TargetGroup {
-                arn: arn.clone(),
-                name: name.clone(),
-                protocol,
-                port,
-                vpc_id,
-                target_type,
-                ip_address_type,
-                protocol_version,
-                health_check_protocol: props
-                    .get("HealthCheckProtocol")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string()),
-                health_check_port: props
-                    .get("HealthCheckPort")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string()),
-                health_check_enabled: props
-                    .get("HealthCheckEnabled")
-                    .and_then(|v| v.as_bool())
-                    .unwrap_or(true),
-                health_check_path: props
-                    .get("HealthCheckPath")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string()),
-                health_check_interval_seconds: props
-                    .get("HealthCheckIntervalSeconds")
-                    .and_then(|v| v.as_i64())
-                    .unwrap_or(30) as i32,
-                health_check_timeout_seconds: props
-                    .get("HealthCheckTimeoutSeconds")
-                    .and_then(|v| v.as_i64())
-                    .unwrap_or(5) as i32,
-                healthy_threshold_count: props
-                    .get("HealthyThresholdCount")
-                    .and_then(|v| v.as_i64())
-                    .unwrap_or(5) as i32,
-                unhealthy_threshold_count: props
-                    .get("UnhealthyThresholdCount")
-                    .and_then(|v| v.as_i64())
-                    .unwrap_or(2) as i32,
-                matcher_http_code: props
-                    .get("Matcher")
-                    .and_then(|v| v.get("HttpCode"))
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string()),
-                matcher_grpc_code: props
-                    .get("Matcher")
-                    .and_then(|v| v.get("GrpcCode"))
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string()),
-                load_balancer_arns: Vec::new(),
-                targets: Vec::new(),
-                tags,
-                attributes: BTreeMap::new(),
-                created_time: Utc::now(),
-            },
-        );
-
-        Ok(ProvisionResult::new(arn.clone())
-            .with("TargetGroupArn", arn)
-            .with("TargetGroupName", name)
-            .with("TargetGroupFullName", format!("targetgroup/{}", &id[..16])))
-    }
-
-    fn delete_elbv2_target_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.target_groups.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_elbv2_listener(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let load_balancer_arn = props
-            .get("LoadBalancerArn")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "LoadBalancerArn is required".to_string())?
-            .to_string();
-        let port = props.get("Port").and_then(|v| v.as_i64()).map(|n| n as i32);
-        let protocol = props
-            .get("Protocol")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let default_actions = parse_elb_actions(props.get("DefaultActions"));
-
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        if !state.load_balancers.contains_key(&load_balancer_arn) {
-            return Err(format!(
-                "LoadBalancer {load_balancer_arn} not yet provisioned"
-            ));
-        }
-
-        let lb_full = load_balancer_arn
-            .rsplit("loadbalancer/")
-            .next()
-            .unwrap_or("")
-            .to_string();
-        let listener_id = Uuid::new_v4().simple().to_string();
-        let arn = format!(
-            "arn:aws:elasticloadbalancing:{}:{}:listener/{}/{}",
-            self.region,
-            self.account_id,
-            lb_full,
-            &listener_id[..16]
-        );
-
-        // Wire forward target groups -> LB association so dataplane probing
-        // and DescribeTargetGroups round-trip the relationship.
-        for action in &default_actions {
-            if let Some(tg_arn) = &action.target_group_arn {
-                if let Some(tg) = state.target_groups.get_mut(tg_arn) {
-                    if !tg.load_balancer_arns.contains(&load_balancer_arn) {
-                        tg.load_balancer_arns.push(load_balancer_arn.clone());
-                    }
-                }
-            }
-            if let Some(forward) = &action.forward {
-                for tgt in &forward.target_groups {
-                    if let Some(tg) = state.target_groups.get_mut(&tgt.target_group_arn) {
-                        if !tg.load_balancer_arns.contains(&load_balancer_arn) {
-                            tg.load_balancer_arns.push(load_balancer_arn.clone());
-                        }
-                    }
-                }
-            }
-        }
-
-        state.listeners.insert(
-            arn.clone(),
-            Listener {
-                arn: arn.clone(),
-                load_balancer_arn,
-                port,
-                protocol,
-                certificates: Vec::new(),
-                ssl_policy: props
-                    .get("SslPolicy")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string()),
-                default_actions,
-                alpn_policy: Vec::new(),
-                mutual_authentication: None,
-                tags: parse_elb_tags(props.get("Tags")),
-                attributes: BTreeMap::new(),
-            },
-        );
-
-        Ok(ProvisionResult::new(arn.clone()).with("ListenerArn", arn))
-    }
-
-    fn delete_elbv2_listener(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.listeners.remove(physical_id);
-        let rules: Vec<String> = state
-            .rules
-            .iter()
-            .filter(|(_, r)| r.listener_arn == physical_id)
-            .map(|(arn, _)| arn.clone())
-            .collect();
-        for r in rules {
-            state.rules.remove(&r);
-        }
-        Ok(())
-    }
-
-    fn create_elbv2_listener_rule(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let listener_arn = props
-            .get("ListenerArn")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ListenerArn is required".to_string())?
-            .to_string();
-        let priority = props
-            .get("Priority")
-            .map(|v| {
-                if let Some(s) = v.as_str() {
-                    s.to_string()
-                } else if let Some(n) = v.as_i64() {
-                    n.to_string()
-                } else {
-                    "1".to_string()
-                }
-            })
-            .unwrap_or_else(|| "1".to_string());
-        let actions = parse_elb_actions(props.get("Actions"));
-        let conditions = parse_elb_rule_conditions(props.get("Conditions"));
-
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        if !state.listeners.contains_key(&listener_arn) {
-            return Err(format!("Listener {listener_arn} not yet provisioned"));
-        }
-        let listener_full = listener_arn
-            .rsplit("listener/")
-            .next()
-            .unwrap_or("")
-            .to_string();
-        let rule_id = Uuid::new_v4().simple().to_string();
-        let arn = format!(
-            "arn:aws:elasticloadbalancing:{}:{}:listener-rule/{}/{}",
-            self.region,
-            self.account_id,
-            listener_full,
-            &rule_id[..16]
-        );
-
-        state.rules.insert(
-            arn.clone(),
-            ElbRule {
-                arn: arn.clone(),
-                listener_arn,
-                priority,
-                conditions,
-                actions,
-                is_default: false,
-                tags: parse_elb_tags(props.get("Tags")),
-            },
-        );
-
-        Ok(ProvisionResult::new(arn.clone()).with("RuleArn", arn))
-    }
-
-    fn delete_elbv2_listener_rule(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.rules.remove(physical_id);
-        Ok(())
-    }
-
-    /// Provision an `AWS::ElasticLoadBalancingV2::ListenerCertificate`.
-    /// Appends each non-default certificate from `Certificates` to the
-    /// target listener (the default listener cert is set on Listener
-    /// creation, so this resource only manages SNI extras).
-    fn create_elbv2_listener_certificate(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let listener_arn = props
-            .get("ListenerArn")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ListenerArn is required".to_string())?
-            .to_string();
-        let certs: Vec<String> = props
-            .get("Certificates")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|c| c.get("CertificateArn").and_then(|v| v.as_str()))
-                    .map(|s| s.to_string())
-                    .collect()
-            })
-            .unwrap_or_default();
-        if certs.is_empty() {
-            return Err("Certificates must contain at least one CertificateArn".to_string());
-        }
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let listener = state
-            .listeners
-            .get_mut(&listener_arn)
-            .ok_or_else(|| format!("Listener {listener_arn} does not exist"))?;
-        for arn in &certs {
-            listener.certificates.retain(|c| &c.certificate_arn != arn);
-            listener.certificates.push(fakecloud_elbv2::Certificate {
-                certificate_arn: arn.clone(),
-                is_default: false,
-            });
-        }
-        Ok(ProvisionResult::new(format!(
-            "{}#{}",
-            listener_arn,
-            certs.join(",")
-        )))
-    }
-
-    fn delete_elbv2_listener_certificate(&self, physical_id: &str) -> Result<(), String> {
-        let (listener_arn, cert_list) = match physical_id.split_once('#') {
-            Some(parts) => parts,
-            None => return Ok(()),
-        };
-        let cert_arns: Vec<&str> = cert_list.split(',').collect();
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        if let Some(listener) = state.listeners.get_mut(listener_arn) {
-            listener
-                .certificates
-                .retain(|c| !cert_arns.iter().any(|a| *a == c.certificate_arn));
-        }
-        Ok(())
-    }
-
-    /// Provision an `AWS::ElasticLoadBalancingV2::TrustStore`.
-    fn create_elbv2_trust_store(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let bucket = props
-            .get("CaCertificatesBundleS3Bucket")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "CaCertificatesBundleS3Bucket is required".to_string())?;
-        let key = props
-            .get("CaCertificatesBundleS3Key")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "CaCertificatesBundleS3Key is required".to_string())?;
-        let tags: Vec<fakecloud_elbv2::Tag> = props
-            .get("Tags")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|t| {
-                        let k = t.get("Key").and_then(|v| v.as_str())?;
-                        let val = t.get("Value").and_then(|v| v.as_str()).unwrap_or("");
-                        Some(fakecloud_elbv2::Tag {
-                            key: k.to_string(),
-                            value: val.to_string(),
-                        })
-                    })
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        if state.trust_stores.values().any(|t| t.name == name) {
-            return Err(format!("Trust store {name} already exists"));
-        }
-        let suffix: String = Uuid::new_v4()
-            .simple()
-            .to_string()
-            .chars()
-            .take(16)
-            .collect();
-        let arn = format!(
-            "arn:aws:elasticloadbalancing:{}:{}:truststore/{}/{}",
-            self.region, self.account_id, name, suffix
-        );
-        let ts = fakecloud_elbv2::TrustStore {
-            arn: arn.clone(),
-            name: name.clone(),
-            status: "ACTIVE".to_string(),
-            number_of_ca_certificates: 1,
-            total_revoked_entries: 0,
-            created_time: Utc::now(),
-            ca_certificates_bundle: Some(format!("s3://{bucket}/{key}").into_bytes()),
-            revocations: BTreeMap::new(),
-            next_revocation_id: 1,
-            tags,
-        };
-        state.trust_stores.insert(arn.clone(), ts);
-        Ok(ProvisionResult::new(arn.clone())
-            .with("TrustStoreArn", arn)
-            .with("Name", name)
-            .with("Status", "ACTIVE".to_string()))
-    }
-
-    fn delete_elbv2_trust_store(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.trust_stores.remove(physical_id);
-        Ok(())
-    }
-
-    /// In-place update for AWS::ElasticLoadBalancingV2::LoadBalancer. Name,
-    /// scheme, type and subnet topology are immutable in real AWS — CFN
-    /// would replace the resource. We only mutate fields the SetSubnets /
-    /// SetSecurityGroups / SetIpAddressType APIs would touch.
-    fn update_elbv2_load_balancer(
-        &self,
-        existing: &StackResource,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let arn = existing.physical_id.clone();
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let lb = state
-            .load_balancers
-            .get_mut(&arn)
-            .ok_or_else(|| format!("LoadBalancer {arn} no longer exists"))?;
-        if let Some(arr) = props.get("SecurityGroups").and_then(|v| v.as_array()) {
-            lb.security_groups = arr
-                .iter()
-                .filter_map(|s| s.as_str().map(|s| s.to_string()))
-                .collect();
-        }
-        if let Some(s) = props.get("IpAddressType").and_then(|v| v.as_str()) {
-            lb.ip_address_type = s.to_string();
-        }
-        if let Some(arr) = props.get("Subnets").and_then(|v| v.as_array()) {
-            let mut zones: Vec<fakecloud_elbv2::AvailabilityZone> = Vec::new();
-            for s in arr {
-                if let Some(subnet_id) = s.as_str() {
-                    zones.push(fakecloud_elbv2::AvailabilityZone {
-                        zone_name: format!("{}a", self.region),
-                        subnet_id: subnet_id.to_string(),
-                        outpost_id: None,
-                        load_balancer_addresses: Vec::new(),
-                        source_nat_ipv6_prefixes: Vec::new(),
-                    });
-                }
-            }
-            lb.availability_zones = zones;
-        }
-        if props.get("Tags").is_some() {
-            lb.tags = parse_elb_tags(props.get("Tags"));
-        }
-        let name = lb.name.clone();
-        let dns_name = lb.dns_name.clone();
-        let canonical = lb.canonical_hosted_zone_id.clone();
-        let lb_full = arn.rsplit("loadbalancer/").next().unwrap_or("").to_string();
-        Ok(ProvisionResult::new(arn.clone())
-            .with("LoadBalancerArn", arn)
-            .with("LoadBalancerFullName", lb_full)
-            .with("LoadBalancerName", name)
-            .with("DNSName", dns_name)
-            .with("CanonicalHostedZoneID", canonical))
-    }
-
-    /// In-place update for AWS::ElasticLoadBalancingV2::TargetGroup. Mirrors
-    /// ModifyTargetGroup: only health-check fields and matcher are mutable.
-    fn update_elbv2_target_group(
-        &self,
-        existing: &StackResource,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let arn = existing.physical_id.clone();
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let tg = state
-            .target_groups
-            .get_mut(&arn)
-            .ok_or_else(|| format!("TargetGroup {arn} no longer exists"))?;
-        if let Some(s) = props.get("HealthCheckProtocol").and_then(|v| v.as_str()) {
-            tg.health_check_protocol = Some(s.to_string());
-        }
-        if let Some(s) = props.get("HealthCheckPort").and_then(|v| v.as_str()) {
-            tg.health_check_port = Some(s.to_string());
-        }
-        if let Some(b) = props.get("HealthCheckEnabled").and_then(|v| v.as_bool()) {
-            tg.health_check_enabled = b;
-        }
-        if let Some(s) = props.get("HealthCheckPath").and_then(|v| v.as_str()) {
-            tg.health_check_path = Some(s.to_string());
-        }
-        if let Some(n) = props.get("HealthCheckIntervalSeconds").and_then(cfn_as_i64) {
-            tg.health_check_interval_seconds = n as i32;
-        }
-        if let Some(n) = props.get("HealthCheckTimeoutSeconds").and_then(cfn_as_i64) {
-            tg.health_check_timeout_seconds = n as i32;
-        }
-        if let Some(n) = props.get("HealthyThresholdCount").and_then(cfn_as_i64) {
-            tg.healthy_threshold_count = n as i32;
-        }
-        if let Some(n) = props.get("UnhealthyThresholdCount").and_then(cfn_as_i64) {
-            tg.unhealthy_threshold_count = n as i32;
-        }
-        if let Some(matcher) = props.get("Matcher") {
-            tg.matcher_http_code = matcher
-                .get("HttpCode")
-                .and_then(|v| v.as_str())
-                .map(|s| s.to_string());
-            tg.matcher_grpc_code = matcher
-                .get("GrpcCode")
-                .and_then(|v| v.as_str())
-                .map(|s| s.to_string());
-        }
-        if props.get("Tags").is_some() {
-            tg.tags = parse_elb_tags(props.get("Tags"));
-        }
-        let name = tg.name.clone();
-        let tg_full = arn
-            .rsplit("targetgroup/")
-            .next()
-            .map(|s| format!("targetgroup/{s}"))
-            .unwrap_or_default();
-        Ok(ProvisionResult::new(arn.clone())
-            .with("TargetGroupArn", arn)
-            .with("TargetGroupName", name)
-            .with("TargetGroupFullName", tg_full))
-    }
-
-    /// In-place update for AWS::ElasticLoadBalancingV2::Listener. Mirrors
-    /// ModifyListener: port, protocol, default actions, certs, ssl policy.
-    fn update_elbv2_listener(
-        &self,
-        existing: &StackResource,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let arn = existing.physical_id.clone();
-        let new_default_actions = props
-            .get("DefaultActions")
-            .map(|v| parse_elb_actions(Some(v)));
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let listener = state
-            .listeners
-            .get_mut(&arn)
-            .ok_or_else(|| format!("Listener {arn} no longer exists"))?;
-        if let Some(n) = props.get("Port").and_then(cfn_as_i64) {
-            listener.port = Some(n as i32);
-        }
-        if let Some(s) = props.get("Protocol").and_then(|v| v.as_str()) {
-            listener.protocol = Some(s.to_string());
-        }
-        if let Some(s) = props.get("SslPolicy").and_then(|v| v.as_str()) {
-            listener.ssl_policy = Some(s.to_string());
-        }
-        if let Some(actions) = new_default_actions {
-            listener.default_actions = actions;
-        }
-        if props.get("Tags").is_some() {
-            listener.tags = parse_elb_tags(props.get("Tags"));
-        }
-        Ok(ProvisionResult::new(arn.clone()).with("ListenerArn", arn))
-    }
-
-    /// In-place update for AWS::ElasticLoadBalancingV2::ListenerRule. Mirrors
-    /// ModifyRule + SetRulePriorities.
-    fn update_elbv2_listener_rule(
-        &self,
-        existing: &StackResource,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let arn = existing.physical_id.clone();
-        let new_actions = props.get("Actions").map(|v| parse_elb_actions(Some(v)));
-        let new_conditions = props
-            .get("Conditions")
-            .map(|v| parse_elb_rule_conditions(Some(v)));
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let rule = state
-            .rules
-            .get_mut(&arn)
-            .ok_or_else(|| format!("ListenerRule {arn} no longer exists"))?;
-        if let Some(v) = props.get("Priority") {
-            rule.priority = if let Some(s) = v.as_str() {
-                s.to_string()
-            } else if let Some(n) = v.as_i64() {
-                n.to_string()
-            } else {
-                rule.priority.clone()
-            };
-        }
-        if let Some(actions) = new_actions {
-            rule.actions = actions;
-        }
-        if let Some(conditions) = new_conditions {
-            rule.conditions = conditions;
-        }
-        if props.get("Tags").is_some() {
-            rule.tags = parse_elb_tags(props.get("Tags"));
-        }
-        Ok(ProvisionResult::new(arn.clone()).with("RuleArn", arn))
-    }
-
-    /// In-place update for AWS::ElasticLoadBalancingV2::ListenerCertificate.
-    /// CFN treats this as replace-on-cert-list-change in real AWS, but we
-    /// can rebuild the SNI cert set against the same physical id without
-    /// disrupting the listener.
-    fn update_elbv2_listener_certificate(
-        &self,
-        existing: &StackResource,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let physical_id = existing.physical_id.clone();
-        let listener_arn = props
-            .get("ListenerArn")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string())
-            .or_else(|| physical_id.split_once('#').map(|(l, _)| l.to_string()))
-            .ok_or_else(|| "ListenerArn is required".to_string())?;
-        let new_certs: Vec<String> = props
-            .get("Certificates")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|c| c.get("CertificateArn").and_then(|v| v.as_str()))
-                    .map(|s| s.to_string())
-                    .collect()
-            })
-            .unwrap_or_default();
-        if new_certs.is_empty() {
-            return Err("Certificates must contain at least one CertificateArn".to_string());
-        }
-
-        // Strip the previously-managed certs, then attach the new set.
-        let prev_certs: Vec<String> = physical_id
-            .split_once('#')
-            .map(|(_, list)| list.split(',').map(|s| s.to_string()).collect())
-            .unwrap_or_default();
-
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let listener = state
-            .listeners
-            .get_mut(&listener_arn)
-            .ok_or_else(|| format!("Listener {listener_arn} does not exist"))?;
-        listener
-            .certificates
-            .retain(|c| !prev_certs.iter().any(|p| p == &c.certificate_arn));
-        for arn in &new_certs {
-            listener.certificates.retain(|c| &c.certificate_arn != arn);
-            listener.certificates.push(fakecloud_elbv2::Certificate {
-                certificate_arn: arn.clone(),
-                is_default: false,
-            });
-        }
-        Ok(ProvisionResult::new(format!(
-            "{}#{}",
-            listener_arn,
-            new_certs.join(",")
-        )))
-    }
-
-    /// In-place update for AWS::ElasticLoadBalancingV2::TrustStore. Only the
-    /// CA bundle and tags are mutable; name is immutable in real AWS.
-    fn update_elbv2_trust_store(
-        &self,
-        existing: &StackResource,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let arn = existing.physical_id.clone();
-        let mut accounts = self.elbv2_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let ts = state
-            .trust_stores
-            .get_mut(&arn)
-            .ok_or_else(|| format!("TrustStore {arn} no longer exists"))?;
-        let new_bucket = props
-            .get("CaCertificatesBundleS3Bucket")
-            .and_then(|v| v.as_str());
-        let new_key = props
-            .get("CaCertificatesBundleS3Key")
-            .and_then(|v| v.as_str());
-        if let (Some(b), Some(k)) = (new_bucket, new_key) {
-            ts.ca_certificates_bundle = Some(format!("s3://{b}/{k}").into_bytes());
-        }
-        if let Some(arr) = props.get("Tags").and_then(|v| v.as_array()) {
-            ts.tags = arr
-                .iter()
-                .filter_map(|t| {
-                    let k = t.get("Key").and_then(|v| v.as_str())?;
-                    let v = t.get("Value").and_then(|v| v.as_str()).unwrap_or("");
-                    Some(fakecloud_elbv2::Tag {
-                        key: k.to_string(),
-                        value: v.to_string(),
-                    })
-                })
-                .collect();
-        }
-        let name = ts.name.clone();
-        let status = ts.status.clone();
-        Ok(ProvisionResult::new(arn.clone())
-            .with("TrustStoreArn", arn)
-            .with("Name", name)
-            .with("Status", status))
-    }
-
-    /// Live-state GetAtt fallback for AWS::ElasticLoadBalancingV2::LoadBalancer.
     fn get_att_elbv2_load_balancer(&self, physical_id: &str, attribute: &str) -> Option<String> {
         let mut accounts = self.elbv2_state.write();
         let state = accounts.get_or_create(&self.account_id);
@@ -3618,582 +2712,6 @@ impl ResourceProvisioner {
 
     // --- Organizations ---
 
-    fn create_organization(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let feature_set = props
-            .get("FeatureSet")
-            .and_then(|v| v.as_str())
-            .unwrap_or("ALL")
-            .to_string();
-
-        let mut org = self.organizations_state.write();
-        if org.is_some() {
-            return Err("Organization already exists; only one per fakecloud process".to_string());
-        }
-        let mut state = OrganizationState::bootstrap(&self.account_id);
-        state.feature_set = feature_set;
-        let org_id = state.org_id.clone();
-        let org_arn = state.org_arn.clone();
-        let mgmt_arn = state.management_account_arn.clone();
-        let root_id = state.root_id.clone();
-        *org = Some(state);
-
-        Ok(ProvisionResult::new(org_id.clone())
-            .with("Id", org_id)
-            .with("Arn", org_arn)
-            .with("ManagementAccountArn", mgmt_arn)
-            .with("RootId", root_id))
-    }
-
-    fn delete_organization(&self, _physical_id: &str) -> Result<(), String> {
-        let mut org = self.organizations_state.write();
-        *org = None;
-        Ok(())
-    }
-
-    fn create_organization_unit(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let parent_id = props
-            .get("ParentId")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ParentId is required".to_string())?
-            .to_string();
-
-        let mut org_lock = self.organizations_state.write();
-        let org = org_lock
-            .as_mut()
-            .ok_or_else(|| "Organization not yet created".to_string())?;
-        // Accept root id, OU id, or `Ref`-resolved logical id (we map to root).
-        let resolved_parent_id = if parent_id == org.root_id || org.ous.contains_key(&parent_id) {
-            parent_id
-        } else {
-            return Err(format!("Parent {parent_id} does not exist"));
-        };
-        let id_suffix: String = Uuid::new_v4()
-            .simple()
-            .to_string()
-            .chars()
-            .take(8)
-            .collect();
-        let id = format!("ou-{}-{}", &org.root_id[2..], id_suffix);
-        let arn = format!(
-            "arn:aws:organizations::{}:ou/{}/{}",
-            org.management_account_id, org.org_id, id
-        );
-        org.ous.insert(
-            id.clone(),
-            OrganizationalUnit {
-                id: id.clone(),
-                arn: arn.clone(),
-                name: name.clone(),
-                parent_id: resolved_parent_id,
-            },
-        );
-        Ok(ProvisionResult::new(id.clone())
-            .with("Id", id)
-            .with("Arn", arn)
-            .with("Name", name))
-    }
-
-    fn delete_organization_unit(&self, physical_id: &str) -> Result<(), String> {
-        let mut org_lock = self.organizations_state.write();
-        if let Some(org) = org_lock.as_mut() {
-            org.ous.remove(physical_id);
-            org.attachments.remove(physical_id);
-        }
-        Ok(())
-    }
-
-    /// Provision an `AWS::Organizations::Account`. Mints a new member
-    /// account synchronously (via Organizations state), optionally moves
-    /// it under the first ParentId when supplied, and persists tags.
-    fn create_organization_account(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let email = props
-            .get("Email")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "Email is required".to_string())?
-            .to_string();
-        let name = props
-            .get("AccountName")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "AccountName is required".to_string())?
-            .to_string();
-        let parent_ids: Vec<String> = props
-            .get("ParentIds")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let tags: Vec<(String, String)> = props
-            .get("Tags")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|t| {
-                        let k = t.get("Key").and_then(|v| v.as_str())?;
-                        let val = t.get("Value").and_then(|v| v.as_str()).unwrap_or("");
-                        Some((k.to_string(), val.to_string()))
-                    })
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        let mut org_lock = self.organizations_state.write();
-        let org = org_lock
-            .as_mut()
-            .ok_or_else(|| "Organization not yet created".to_string())?;
-        // CFN provisioning is its own asynchronous flow; we don't need
-        // a second layer of poll-for-completion on top. Begin the
-        // request and immediately drive it to SUCCEEDED so the rest of
-        // this provisioner sees a fully enrolled account.
-        let pending = org.begin_create_account(&email, &name, None);
-        let status = org.complete_create_account(&pending.id).unwrap_or(pending);
-        let account_id = status
-            .account_id
-            .clone()
-            .ok_or_else(|| "create_account did not return an account id".to_string())?;
-        let account_arn = org
-            .accounts
-            .get(&account_id)
-            .map(|a| a.arn.clone())
-            .unwrap_or_default();
-        let joined_method = org
-            .accounts
-            .get(&account_id)
-            .map(|a| a.joined_method.clone())
-            .unwrap_or_else(|| "CREATED".to_string());
-        let joined_timestamp = org
-            .accounts
-            .get(&account_id)
-            .map(|a| a.joined_timestamp.to_rfc3339())
-            .unwrap_or_default();
-        let acct_status = org
-            .accounts
-            .get(&account_id)
-            .map(|a| a.status.clone())
-            .unwrap_or_else(|| "ACTIVE".to_string());
-
-        if let Some(parent) = parent_ids.first() {
-            let source = org
-                .accounts
-                .get(&account_id)
-                .map(|a| a.parent_id.clone())
-                .unwrap_or_else(|| org.root_id.clone());
-            if parent != &source {
-                org.move_account(&account_id, &source, parent)
-                    .map_err(|e| format!("Failed to move account to parent {parent}: {e:?}"))?;
-            }
-        }
-
-        if !tags.is_empty() {
-            org.set_resource_tags(&account_id, &tags);
-        }
-
-        Ok(ProvisionResult::new(account_id.clone())
-            .with("AccountId", account_id)
-            .with("AccountName", name)
-            .with("Email", email)
-            .with("Arn", account_arn)
-            .with("JoinedMethod", joined_method)
-            .with("JoinedTimestamp", joined_timestamp)
-            .with("Status", acct_status))
-    }
-
-    /// Close the member account on stack delete. Real AWS leaves the
-    /// account in `SUSPENDED` for 90 days; we just flip it via
-    /// `close_account` so subsequent reads see it as suspended.
-    fn delete_organization_account(&self, physical_id: &str) -> Result<(), String> {
-        let mut org_lock = self.organizations_state.write();
-        if let Some(org) = org_lock.as_mut() {
-            let _ = org.close_account(physical_id);
-        }
-        Ok(())
-    }
-
-    fn create_organization_policy(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let description = props
-            .get("Description")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let policy_type = props
-            .get("Type")
-            .and_then(|v| v.as_str())
-            .unwrap_or(POLICY_TYPE_SCP)
-            .to_string();
-        let content = props
-            .get("Content")
-            .map(|v| {
-                if v.is_string() {
-                    v.as_str().unwrap_or("").to_string()
-                } else {
-                    serde_json::to_string(v).unwrap_or_default()
-                }
-            })
-            .unwrap_or_default();
-        let target_ids: Vec<String> = props
-            .get("TargetIds")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|t| t.as_str().map(|s| s.to_string()))
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        let mut org_lock = self.organizations_state.write();
-        let org = org_lock
-            .as_mut()
-            .ok_or_else(|| "Organization not yet created".to_string())?;
-        let id_suffix: String = Uuid::new_v4()
-            .simple()
-            .to_string()
-            .chars()
-            .take(8)
-            .collect();
-        let id = format!("p-{}", id_suffix);
-        let arn = format!(
-            "arn:aws:organizations::{}:policy/{}/{}/{}",
-            org.management_account_id,
-            org.org_id,
-            policy_type.to_lowercase(),
-            id
-        );
-        org.policies.insert(
-            id.clone(),
-            OrgPolicy {
-                id: id.clone(),
-                arn: arn.clone(),
-                name: name.clone(),
-                description,
-                policy_type,
-                aws_managed: false,
-                content,
-            },
-        );
-        for target in target_ids {
-            org.attachments
-                .entry(target)
-                .or_default()
-                .insert(id.clone());
-        }
-        Ok(ProvisionResult::new(id.clone())
-            .with("Id", id)
-            .with("Arn", arn)
-            .with("Name", name))
-    }
-
-    fn delete_organization_policy(&self, physical_id: &str) -> Result<(), String> {
-        let mut org_lock = self.organizations_state.write();
-        if let Some(org) = org_lock.as_mut() {
-            org.policies.remove(physical_id);
-            for attachments in org.attachments.values_mut() {
-                attachments.remove(physical_id);
-            }
-        }
-        Ok(())
-    }
-
-    fn create_organization_resource_policy(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let content = props
-            .get("Content")
-            .map(|v| {
-                if v.is_string() {
-                    v.as_str().unwrap_or("").to_string()
-                } else {
-                    serde_json::to_string(v).unwrap_or_default()
-                }
-            })
-            .ok_or_else(|| "Content is required".to_string())?;
-
-        let mut org_lock = self.organizations_state.write();
-        let org = org_lock
-            .as_mut()
-            .ok_or_else(|| "Organization not yet created".to_string())?;
-        org.resource_policy = Some(content);
-        let arn = format!(
-            "arn:aws:organizations::{}:resourcepolicy/{}/rp",
-            org.management_account_id, org.org_id
-        );
-        Ok(ProvisionResult::new(arn.clone()).with("Arn", arn))
-    }
-
-    fn delete_organization_resource_policy(&self, _physical_id: &str) -> Result<(), String> {
-        let mut org_lock = self.organizations_state.write();
-        if let Some(org) = org_lock.as_mut() {
-            org.resource_policy = None;
-        }
-        Ok(())
-    }
-
-    fn delete_log_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.default_mut();
-        // physical_id is the ARN; find the log group name
-        let name = state
-            .log_groups
-            .iter()
-            .find(|(_, g)| g.arn == physical_id)
-            .map(|(name, _)| name.clone());
-        if let Some(name) = name {
-            state.log_groups.remove(&name);
-        }
-        Ok(())
-    }
-
-    fn create_log_stream(&self, resource: &ResourceDefinition) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let log_group_name = props
-            .get("LogGroupName")
-            .and_then(|v| v.as_str())
-            .map(parse_log_group_name)
-            .ok_or_else(|| "LogGroupName is required".to_string())?;
-        let log_stream_name = props
-            .get("LogStreamName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        let group = state
-            .log_groups
-            .get_mut(&log_group_name)
-            .ok_or_else(|| format!("Log group {log_group_name} does not exist"))?;
-        let arn = format!(
-            "arn:aws:logs:{}:{}:log-group:{}:log-stream:{}",
-            self.region, self.account_id, log_group_name, log_stream_name
-        );
-        if group.log_streams.contains_key(&log_stream_name) {
-            return Err(format!(
-                "Log stream {log_stream_name} already exists in {log_group_name}"
-            ));
-        }
-        group.log_streams.insert(
-            log_stream_name.clone(),
-            LogStream {
-                name: log_stream_name.clone(),
-                arn,
-                creation_time: Utc::now().timestamp_millis(),
-                first_event_timestamp: None,
-                last_event_timestamp: None,
-                last_ingestion_time: None,
-                upload_sequence_token: String::new(),
-                events: Vec::new(),
-            },
-        );
-
-        // Encode group + stream into the physical id so deletion can target both.
-        let physical_id = format!("{log_group_name}|{log_stream_name}");
-        Ok(ProvisionResult::new(physical_id))
-    }
-
-    fn delete_log_stream(&self, physical_id: &str) -> Result<(), String> {
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        if let Some((group_name, stream_name)) = physical_id.split_once('|') {
-            if let Some(group) = state.log_groups.get_mut(group_name) {
-                group.log_streams.remove(stream_name);
-            }
-        }
-        Ok(())
-    }
-
-    fn create_metric_filter(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let log_group_name = props
-            .get("LogGroupName")
-            .and_then(|v| v.as_str())
-            .map(parse_log_group_name)
-            .ok_or_else(|| "LogGroupName is required".to_string())?;
-        let filter_name = props
-            .get("FilterName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let filter_pattern = props
-            .get("FilterPattern")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-
-        let mut transformations: Vec<MetricTransformation> = Vec::new();
-        if let Some(arr) = props
-            .get("MetricTransformations")
-            .and_then(|v| v.as_array())
-        {
-            for t in arr {
-                let metric_name = t
-                    .get("MetricName")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let metric_namespace = t
-                    .get("MetricNamespace")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let metric_value = t
-                    .get("MetricValue")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("1")
-                    .to_string();
-                let default_value = t.get("DefaultValue").and_then(|v| v.as_f64());
-                let unit = t
-                    .get("Unit")
-                    .and_then(|v| v.as_str())
-                    .map(|s| s.to_string());
-                transformations.push(MetricTransformation {
-                    metric_name,
-                    metric_namespace,
-                    metric_value,
-                    default_value,
-                    unit,
-                });
-            }
-        }
-
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        if !state.log_groups.contains_key(&log_group_name) {
-            return Err(format!("Log group {log_group_name} does not exist"));
-        }
-        state
-            .metric_filters
-            .retain(|f| !(f.log_group_name == log_group_name && f.filter_name == filter_name));
-        state.metric_filters.push(MetricFilter {
-            filter_name: filter_name.clone(),
-            filter_pattern,
-            log_group_name: log_group_name.clone(),
-            metric_transformations: transformations,
-            creation_time: Utc::now().timestamp_millis(),
-        });
-
-        Ok(ProvisionResult::new(format!(
-            "{log_group_name}|{filter_name}"
-        )))
-    }
-
-    fn delete_metric_filter(&self, physical_id: &str) -> Result<(), String> {
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        if let Some((group_name, filter_name)) = physical_id.split_once('|') {
-            state
-                .metric_filters
-                .retain(|f| !(f.log_group_name == group_name && f.filter_name == filter_name));
-        }
-        Ok(())
-    }
-
-    fn create_subscription_filter(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let log_group_name = props
-            .get("LogGroupName")
-            .and_then(|v| v.as_str())
-            .map(parse_log_group_name)
-            .ok_or_else(|| "LogGroupName is required".to_string())?;
-        let filter_name = props
-            .get("FilterName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let filter_pattern = props
-            .get("FilterPattern")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let destination_arn = props
-            .get("DestinationArn")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "DestinationArn is required".to_string())?
-            .to_string();
-        let role_arn = props
-            .get("RoleArn")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let distribution = props
-            .get("Distribution")
-            .and_then(|v| v.as_str())
-            .unwrap_or("ByLogStream")
-            .to_string();
-
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        let group = state
-            .log_groups
-            .get_mut(&log_group_name)
-            .ok_or_else(|| format!("Log group {log_group_name} does not exist"))?;
-        group
-            .subscription_filters
-            .retain(|f| f.filter_name != filter_name);
-        group.subscription_filters.push(SubscriptionFilter {
-            filter_name: filter_name.clone(),
-            log_group_name: log_group_name.clone(),
-            filter_pattern,
-            destination_arn,
-            role_arn,
-            distribution,
-            creation_time: Utc::now().timestamp_millis(),
-        });
-
-        Ok(ProvisionResult::new(format!(
-            "{log_group_name}|{filter_name}"
-        )))
-    }
-
-    fn delete_subscription_filter(&self, physical_id: &str) -> Result<(), String> {
-        let mut logs_accounts = self.logs_state.write();
-        let state = logs_accounts.get_or_create(&self.account_id);
-        if let Some((group_name, filter_name)) = physical_id.split_once('|') {
-            if let Some(group) = state.log_groups.get_mut(group_name) {
-                group
-                    .subscription_filters
-                    .retain(|f| f.filter_name != filter_name);
-            }
-        }
-        Ok(())
-    }
-
-    // --- Custom Resources ---
-
-    /// Invoke a Lambda function synchronously via the delivery bus.
     fn invoke_lambda_sync(&self, function_arn: &str, payload: &str) -> Result<(), String> {
         let delivery = self.delivery.clone();
         let function_arn = function_arn.to_string();
@@ -4316,1651 +2834,6 @@ impl ResourceProvisioner {
     }
 
     // --- Application Auto Scaling ---
-
-    fn create_application_autoscaling_scalable_target(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let service_namespace = props
-            .get("ServiceNamespace")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ServiceNamespace is required".to_string())?
-            .to_string();
-        let resource_id = props
-            .get("ResourceId")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ResourceId is required".to_string())?
-            .to_string();
-        let scalable_dimension = props
-            .get("ScalableDimension")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ScalableDimension is required".to_string())?
-            .to_string();
-        let min_capacity = props
-            .get("MinCapacity")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .ok_or_else(|| "MinCapacity is required".to_string())?;
-        let max_capacity = props
-            .get("MaxCapacity")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .ok_or_else(|| "MaxCapacity is required".to_string())?;
-        if min_capacity > max_capacity {
-            return Err("MinCapacity must be <= MaxCapacity".to_string());
-        }
-        let role_arn = props
-            .get("RoleARN")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let suspended_state = props.get("SuspendedState").map(|v| AppasSuspendedState {
-            dynamic_scaling_in_suspended: v
-                .get("DynamicScalingInSuspended")
-                .and_then(|x| x.as_bool()),
-            dynamic_scaling_out_suspended: v
-                .get("DynamicScalingOutSuspended")
-                .and_then(|x| x.as_bool()),
-            scheduled_scaling_suspended: v
-                .get("ScheduledScalingSuspended")
-                .and_then(|x| x.as_bool()),
-        });
-
-        let arn = format!(
-            "arn:aws:application-autoscaling:{}:{}:scalable-target/{}",
-            self.region,
-            self.account_id,
-            fakecloud_core::ids::short_id(10)
-        );
-        let role = role_arn.unwrap_or_else(|| {
-            let suffix = match service_namespace.as_str() {
-                "ecs" => "ECSService",
-                "elasticmapreduce" => "EMRContainerService",
-                "ec2" => "EC2SpotFleetRequest",
-                "appstream" => "ApplicationAutoScaling_AppStreamFleet",
-                "dynamodb" => "DynamoDBTable",
-                "rds" => "RDSCluster",
-                "sagemaker" => "SageMakerEndpoint",
-                "lambda" => "LambdaConcurrency",
-                "elasticache" => "ElastiCacheRG",
-                "cassandra" => "CassandraTable",
-                "kafka" => "KafkaCluster",
-                _ => "ApplicationAutoScaling_Default",
-            };
-            format!(
-                "arn:aws:iam::{}:role/aws-service-role/applicationautoscaling.amazonaws.com/AWSServiceRoleForApplicationAutoScaling_{}",
-                self.account_id, suffix
-            )
-        });
-
-        let mut state = self.app_autoscaling_state.write();
-        let account = state.accounts.entry(self.account_id.clone()).or_default();
-        let key = (
-            service_namespace.clone(),
-            resource_id.clone(),
-            scalable_dimension.clone(),
-        );
-        let target = AppasScalableTarget {
-            arn: arn.clone(),
-            service_namespace: service_namespace.clone(),
-            resource_id: resource_id.clone(),
-            scalable_dimension: scalable_dimension.clone(),
-            min_capacity,
-            max_capacity,
-            role_arn: role,
-            creation_time: Utc::now(),
-            suspended_state,
-            predicted_capacity: None,
-        };
-        account.scalable_targets.insert(key, target);
-
-        Ok(ProvisionResult::new(resource_id.clone())
-            .with("ScalableTargetARN", arn)
-            .with("ServiceNamespace", service_namespace)
-            .with("ScalableDimension", scalable_dimension))
-    }
-
-    fn create_application_autoscaling_scaling_policy(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let policy_name = props
-            .get("PolicyName")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "PolicyName is required".to_string())?
-            .to_string();
-        let service_namespace = props
-            .get("ServiceNamespace")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ServiceNamespace is required".to_string())?
-            .to_string();
-        let resource_id = props
-            .get("ResourceId")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ResourceId is required".to_string())?
-            .to_string();
-        let scalable_dimension = props
-            .get("ScalableDimension")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "ScalableDimension is required".to_string())?
-            .to_string();
-        let policy_type = props
-            .get("PolicyType")
-            .and_then(|v| v.as_str())
-            .unwrap_or("StepScaling")
-            .to_string();
-        let step_cfg = props.get("StepScalingPolicyConfiguration").cloned();
-        let tt_cfg = props
-            .get("TargetTrackingScalingPolicyConfiguration")
-            .cloned();
-        let pred_cfg = props.get("PredictiveScalingPolicyConfiguration").cloned();
-
-        let target_key = (
-            service_namespace.clone(),
-            resource_id.clone(),
-            scalable_dimension.clone(),
-        );
-        let policy_key = (
-            service_namespace.clone(),
-            resource_id.clone(),
-            scalable_dimension.clone(),
-            policy_name.clone(),
-        );
-
-        let mut state = self.app_autoscaling_state.write();
-        let account = state.accounts.entry(self.account_id.clone()).or_default();
-        if !account.scalable_targets.contains_key(&target_key) {
-            return Err(format!(
-                "No scalable target registered for ServiceNamespace={} ResourceId={} ScalableDimension={}",
-                service_namespace, resource_id, scalable_dimension
-            ));
-        }
-        let arn = format!(
-            "arn:aws:autoscaling:{}:{}:scalingPolicy:{}:resource/{}/{}:policyName/{}",
-            self.region,
-            self.account_id,
-            Uuid::new_v4(),
-            service_namespace,
-            resource_id,
-            policy_name
-        );
-        let policy = AppasScalingPolicy {
-            arn: arn.clone(),
-            policy_name: policy_name.clone(),
-            service_namespace: service_namespace.clone(),
-            resource_id: resource_id.clone(),
-            scalable_dimension: scalable_dimension.clone(),
-            policy_type: policy_type.clone(),
-            creation_time: Utc::now(),
-            step_scaling_policy_configuration: step_cfg,
-            target_tracking_scaling_policy_configuration: tt_cfg,
-            predictive_scaling_policy_configuration: pred_cfg,
-            alarms: Vec::new(),
-            last_applied_at: None,
-        };
-        account.scaling_policies.insert(policy_key, policy);
-
-        Ok(ProvisionResult::new(arn.clone())
-            .with("PolicyName", policy_name)
-            .with("ServiceNamespace", service_namespace)
-            .with("ResourceId", resource_id)
-            .with("ScalableDimension", scalable_dimension))
-    }
-
-    fn delete_application_autoscaling_scalable_target(
-        &self,
-        physical_id: &str,
-        attributes: &BTreeMap<String, String>,
-    ) -> Result<(), String> {
-        let namespace = attributes
-            .get("ServiceNamespace")
-            .cloned()
-            .ok_or_else(|| "ServiceNamespace missing in attributes".to_string())?;
-        let resource_id = physical_id.to_string();
-        let dimension = attributes
-            .get("ScalableDimension")
-            .cloned()
-            .ok_or_else(|| "ScalableDimension missing in attributes".to_string())?;
-        let key = (namespace, resource_id.clone(), dimension);
-
-        let mut state = self.app_autoscaling_state.write();
-        let account = state.accounts.entry(self.account_id.clone()).or_default();
-        account.scalable_targets.remove(&key);
-        account
-            .scaling_policies
-            .retain(|k, _| !(k.0 == key.0 && k.1 == key.1 && k.2 == key.2));
-        account
-            .scheduled_actions
-            .retain(|k, _| !(k.0 == key.0 && k.1 == key.1 && k.2 == key.2));
-        Ok(())
-    }
-
-    fn delete_application_autoscaling_scaling_policy(
-        &self,
-        _physical_id: &str,
-        attributes: &BTreeMap<String, String>,
-    ) -> Result<(), String> {
-        let policy_name = attributes
-            .get("PolicyName")
-            .cloned()
-            .ok_or_else(|| "PolicyName missing in attributes".to_string())?;
-        let namespace = attributes
-            .get("ServiceNamespace")
-            .cloned()
-            .ok_or_else(|| "ServiceNamespace missing in attributes".to_string())?;
-        let resource_id = attributes
-            .get("ResourceId")
-            .cloned()
-            .ok_or_else(|| "ResourceId missing in attributes".to_string())?;
-        let dimension = attributes
-            .get("ScalableDimension")
-            .cloned()
-            .ok_or_else(|| "ScalableDimension missing in attributes".to_string())?;
-        let key = (namespace, resource_id, dimension, policy_name);
-
-        let mut state = self.app_autoscaling_state.write();
-        let account = state.accounts.entry(self.account_id.clone()).or_default();
-        account.scaling_policies.remove(&key);
-        Ok(())
-    }
-
-    // --- ElastiCache ---
-
-    fn create_ec_parameter_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("CacheParameterGroupName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let family = props
-            .get("CacheParameterGroupFamily")
-            .and_then(|v| v.as_str())
-            .unwrap_or("redis7")
-            .to_string();
-        let description = props
-            .get("Description")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:parametergroup:{}",
-            self.region, self.account_id, name
-        );
-        let group = CacheParameterGroup {
-            cache_parameter_group_name: name.clone(),
-            cache_parameter_group_family: family,
-            description,
-            is_global: false,
-            arn: arn.clone(),
-        };
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        // ParameterGroups stored as Vec — replace any existing entry.
-        state
-            .parameter_groups
-            .retain(|p| p.cache_parameter_group_name != name);
-        state.parameter_groups.push(group);
-        Ok(ProvisionResult::new(name).with("Arn", arn))
-    }
-
-    fn delete_ec_parameter_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state
-            .parameter_groups
-            .retain(|p| p.cache_parameter_group_name != physical_id);
-        Ok(())
-    }
-
-    fn create_ec_subnet_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("CacheSubnetGroupName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let description = props
-            .get("Description")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let subnet_ids: Vec<String> = props
-            .get("SubnetIds")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:subnetgroup:{}",
-            self.region, self.account_id, name
-        );
-        let group = CacheSubnetGroup {
-            cache_subnet_group_name: name.clone(),
-            cache_subnet_group_description: description,
-            vpc_id: String::new(),
-            subnet_ids,
-            arn: arn.clone(),
-        };
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.subnet_groups.insert(name.clone(), group);
-        Ok(ProvisionResult::new(name).with("Arn", arn))
-    }
-
-    fn delete_ec_subnet_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.subnet_groups.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_ec_security_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("CacheSecurityGroupName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let description = props
-            .get("Description")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:securitygroup:{}",
-            self.region, self.account_id, name
-        );
-        let group = CacheSecurityGroup {
-            cache_security_group_name: name.clone(),
-            description,
-            owner_id: self.account_id.clone(),
-            arn: arn.clone(),
-            ec2_security_groups: Vec::new(),
-        };
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.security_groups.insert(name.clone(), group);
-        Ok(ProvisionResult::new(name).with("Arn", arn))
-    }
-
-    fn delete_ec_security_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.security_groups.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_ec_user(&self, resource: &ResourceDefinition) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let user_id = props
-            .get("UserId")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let user_name = props
-            .get("UserName")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&user_id)
-            .to_string();
-        let engine = props
-            .get("Engine")
-            .and_then(|v| v.as_str())
-            .unwrap_or("redis")
-            .to_string();
-        let access_string = props
-            .get("AccessString")
-            .and_then(|v| v.as_str())
-            .unwrap_or("on ~* +@all")
-            .to_string();
-        let authentication_type = props
-            .get("AuthenticationMode")
-            .and_then(|v| v.get("Type"))
-            .and_then(|v| v.as_str())
-            .unwrap_or("no-password-required")
-            .to_string();
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:user:{}",
-            self.region, self.account_id, user_id
-        );
-        let user = EcUser {
-            user_id: user_id.clone(),
-            user_name,
-            engine,
-            access_string,
-            status: "active".to_string(),
-            authentication_type,
-            password_count: 0,
-            arn: arn.clone(),
-            minimum_engine_version: "6.0".to_string(),
-            user_group_ids: Vec::new(),
-        };
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.users.insert(user_id.clone(), user);
-        Ok(ProvisionResult::new(user_id).with("Arn", arn))
-    }
-
-    fn delete_ec_user(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.users.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_ec_user_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let user_group_id = props
-            .get("UserGroupId")
-            .and_then(|v| v.as_str())
-            .unwrap_or(&resource.logical_id)
-            .to_string();
-        let engine = props
-            .get("Engine")
-            .and_then(|v| v.as_str())
-            .unwrap_or("redis")
-            .to_string();
-        let user_ids: Vec<String> = props
-            .get("UserIds")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:usergroup:{}",
-            self.region, self.account_id, user_group_id
-        );
-        let group = EcUserGroup {
-            user_group_id: user_group_id.clone(),
-            engine,
-            status: "active".to_string(),
-            user_ids,
-            arn: arn.clone(),
-            minimum_engine_version: "6.0".to_string(),
-            pending_changes: None,
-            replication_groups: Vec::new(),
-        };
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.user_groups.insert(user_group_id.clone(), group);
-        Ok(ProvisionResult::new(user_group_id).with("Arn", arn))
-    }
-
-    fn delete_ec_user_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        state.user_groups.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_ec_cache_cluster(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let id = props
-            .get("ClusterName")
-            .and_then(|v| v.as_str())
-            .map(String::from)
-            .unwrap_or_else(|| format!("cfn-cc-{}", resource.logical_id.to_lowercase()));
-        let cache_node_type = props
-            .get("CacheNodeType")
-            .and_then(|v| v.as_str())
-            .unwrap_or("cache.t4g.micro")
-            .to_string();
-        let engine = props
-            .get("Engine")
-            .and_then(|v| v.as_str())
-            .unwrap_or("redis")
-            .to_string();
-        let engine_version = props
-            .get("EngineVersion")
-            .and_then(|v| v.as_str())
-            .unwrap_or("7.1")
-            .to_string();
-        let num_cache_nodes = props
-            .get("NumCacheNodes")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .unwrap_or(1);
-        let preferred_az = props
-            .get("PreferredAvailabilityZone")
-            .and_then(|v| v.as_str())
-            .unwrap_or("us-east-1a")
-            .to_string();
-        let cache_subnet_group_name = props
-            .get("CacheSubnetGroupName")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let auto_minor_version_upgrade = props
-            .get("AutoMinorVersionUpgrade")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(true);
-        let default_port = if engine == "memcached" { 11211 } else { 6379 };
-        let port = props
-            .get("Port")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as u16)
-            .unwrap_or(default_port);
-        let cache_parameter_group_name = props
-            .get("CacheParameterGroupName")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let security_group_ids: Vec<String> = props
-            .get("VpcSecurityGroupIds")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let cache_security_group_names: Vec<String> = props
-            .get("CacheSecurityGroupNames")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let preferred_availability_zones: Vec<String> = props
-            .get("PreferredAvailabilityZones")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let snapshot_arns: Vec<String> = props
-            .get("SnapshotArns")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let snapshot_name = props
-            .get("SnapshotName")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let snapshot_retention_limit = props
-            .get("SnapshotRetentionLimit")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .unwrap_or(0);
-        let snapshot_window = props
-            .get("SnapshotWindow")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let preferred_maintenance_window = props
-            .get("PreferredMaintenanceWindow")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let notification_topic_arn = props
-            .get("NotificationTopicArn")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let transit_encryption_enabled = props
-            .get("TransitEncryptionEnabled")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let auth_token = props
-            .get("AuthToken")
-            .and_then(|v| v.as_str())
-            .filter(|s| !s.is_empty())
-            .map(String::from);
-        let auth_token_enabled = auth_token.is_some();
-        let network_type = props
-            .get("NetworkType")
-            .and_then(|v| v.as_str())
-            .map(String::from)
-            .or_else(|| Some("ipv4".to_string()));
-        let ip_discovery = props
-            .get("IpDiscovery")
-            .and_then(|v| v.as_str())
-            .map(String::from)
-            .or_else(|| Some("ipv4".to_string()));
-        let az_mode = props
-            .get("AZMode")
-            .and_then(|v| v.as_str())
-            .map(String::from)
-            .or_else(|| Some("single-az".to_string()));
-        let outpost_mode = props
-            .get("OutpostMode")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let preferred_outpost_arn = props
-            .get("PreferredOutpostArn")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        // When an ElastiCache container runtime is configured, the record is
-        // inserted as "creating" and `CreateStack` drains a spawn intent that
-        // backs it with a real Redis/Memcached container (flipping it to
-        // "available" once up), matching the direct `CreateCacheCluster` path.
-        // Without a runtime (CI / metadata-only) it stays "available" with no
-        // container, as before.
-        let back_with_container = self.elasticache_runtime.is_some();
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:cluster:{}",
-            state.region, state.account_id, id
-        );
-        let endpoint_address = format!("{id}.fakecloud.{}.cache.amazonaws.com", state.region);
-        let cluster = EcCacheCluster {
-            cache_cluster_id: id.clone(),
-            cache_node_type,
-            engine,
-            engine_version,
-            cache_cluster_status: if back_with_container {
-                "creating".to_string()
-            } else {
-                "available".to_string()
-            },
-            num_cache_nodes,
-            preferred_availability_zone: preferred_az,
-            cache_subnet_group_name,
-            auto_minor_version_upgrade,
-            arn: arn.clone(),
-            created_at: Utc::now().to_rfc3339(),
-            endpoint_address: endpoint_address.clone(),
-            endpoint_port: port,
-            container_id: String::new(),
-            host_port: 0,
-            replication_group_id: None,
-            cache_parameter_group_name,
-            security_group_ids,
-            log_delivery_configurations: Vec::new(),
-            transit_encryption_enabled,
-            at_rest_encryption_enabled: false,
-            auth_token_enabled,
-            port,
-            preferred_maintenance_window,
-            preferred_availability_zones,
-            notification_topic_arn,
-            cache_security_group_names,
-            snapshot_arns,
-            snapshot_name,
-            snapshot_retention_limit,
-            snapshot_window,
-            outpost_mode,
-            preferred_outpost_arn,
-            network_type,
-            ip_discovery,
-            az_mode,
-            auth_token,
-            kms_key_id: None,
-            transit_encryption_mode: None,
-            data_tiering_enabled: None,
-            cluster_mode: None,
-            preferred_outpost_arns: Vec::new(),
-        };
-        state.cache_clusters.insert(id.clone(), cluster);
-        drop(accounts);
-
-        if back_with_container {
-            self.pending_container_spawns
-                .lock()
-                .push(ContainerSpawnIntent::ElastiCacheCluster {
-                    cache_cluster_id: id.clone(),
-                });
-        }
-
-        Ok(ProvisionResult::new(id.clone())
-            .with("Arn", arn)
-            .with("RedisEndpoint.Address", endpoint_address.clone())
-            .with("RedisEndpoint.Port", port.to_string())
-            .with("ConfigurationEndpoint.Address", endpoint_address)
-            .with("ConfigurationEndpoint.Port", port.to_string()))
-    }
-
-    fn delete_ec_cache_cluster(&self, physical_id: &str) -> Result<(), String> {
-        {
-            let mut accounts = self.elasticache_state.write();
-            let state = accounts.get_or_create(&self.account_id);
-            state.cache_clusters.remove(physical_id);
-        }
-        if self.elasticache_runtime.is_some() {
-            self.pending_container_teardowns.lock().push(
-                ContainerTeardownIntent::ElastiCacheCluster {
-                    cache_cluster_id: physical_id.to_string(),
-                },
-            );
-        }
-        Ok(())
-    }
-
-    fn create_ec_replication_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let id = props
-            .get("ReplicationGroupId")
-            .and_then(|v| v.as_str())
-            .map(String::from)
-            .unwrap_or_else(|| format!("cfn-rg-{}", resource.logical_id.to_lowercase()));
-        let description = props
-            .get("ReplicationGroupDescription")
-            .and_then(|v| v.as_str())
-            .unwrap_or("CFN-provisioned replication group")
-            .to_string();
-        let cache_node_type = props
-            .get("CacheNodeType")
-            .and_then(|v| v.as_str())
-            .unwrap_or("cache.t4g.micro")
-            .to_string();
-        let engine = props
-            .get("Engine")
-            .and_then(|v| v.as_str())
-            .unwrap_or("redis")
-            .to_string();
-        let engine_version = props
-            .get("EngineVersion")
-            .and_then(|v| v.as_str())
-            .unwrap_or("7.1")
-            .to_string();
-        let num_cache_clusters = props
-            .get("NumCacheClusters")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .unwrap_or(1);
-        let num_node_groups = props
-            .get("NumNodeGroups")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .unwrap_or(0);
-        let replicas_per_node_group = props
-            .get("ReplicasPerNodeGroup")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32);
-        let automatic_failover_enabled = props
-            .get("AutomaticFailoverEnabled")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let multi_az_enabled = props
-            .get("MultiAZEnabled")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let transit_encryption_enabled = props
-            .get("TransitEncryptionEnabled")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let at_rest_encryption_enabled = props
-            .get("AtRestEncryptionEnabled")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let kms_key_id = props
-            .get("KmsKeyId")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let auth_token_enabled = props
-            .get("AuthToken")
-            .and_then(|v| v.as_str())
-            .filter(|s| !s.is_empty())
-            .is_some();
-        let user_group_ids: Vec<String> = props
-            .get("UserGroupIds")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let snapshot_retention_limit = props
-            .get("SnapshotRetentionLimit")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as i32)
-            .unwrap_or(0);
-        let snapshot_window = props
-            .get("SnapshotWindow")
-            .and_then(|v| v.as_str())
-            .unwrap_or("00:00-01:00")
-            .to_string();
-        let port = props
-            .get("Port")
-            .and_then(|v| v.as_i64())
-            .map(|n| n as u16)
-            .unwrap_or(6379);
-        let cluster_enabled = num_node_groups > 1
-            || props
-                .get("ClusterEnabled")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(false);
-
-        // As with CacheCluster: with a runtime configured the group is inserted
-        // as "creating" and `CreateStack` drains a spawn intent that backs it
-        // with a real Redis container, matching the direct
-        // `CreateReplicationGroup` path. Without a runtime it stays "available"
-        // with no container.
-        let back_with_container = self.elasticache_runtime.is_some();
-        let mut accounts = self.elasticache_state.write();
-        let state = accounts.get_or_create(&self.account_id);
-        let arn = format!(
-            "arn:aws:elasticache:{}:{}:replicationgroup:{}",
-            state.region, state.account_id, id
-        );
-        let endpoint_address = format!(
-            "{id}.fakecloud.ng.0001.{}.cache.amazonaws.com",
-            state.region
-        );
-        let configuration_endpoint = if cluster_enabled {
-            Some(format!(
-                "{id}.fakecloud.cfg.{}.cache.amazonaws.com",
-                state.region
-            ))
-        } else {
-            None
-        };
-
-        let group = EcReplicationGroup {
-            replication_group_id: id.clone(),
-            description,
-            global_replication_group_id: None,
-            global_replication_group_role: None,
-            status: if back_with_container {
-                "creating".to_string()
-            } else {
-                "available".to_string()
-            },
-            cache_node_type,
-            engine,
-            engine_version,
-            num_cache_clusters,
-            automatic_failover_enabled,
-            endpoint_address: endpoint_address.clone(),
-            endpoint_port: port,
-            arn: arn.clone(),
-            created_at: Utc::now().to_rfc3339(),
-            container_id: String::new(),
-            host_port: 0,
-            member_clusters: Vec::new(),
-            snapshot_retention_limit,
-            snapshot_window,
-            transit_encryption_enabled,
-            at_rest_encryption_enabled,
-            cluster_enabled,
-            kms_key_id,
-            auth_token_enabled,
-            user_group_ids,
-            multi_az_enabled,
-            log_delivery_configurations: Vec::new(),
-            data_tiering: props
-                .get("DataTieringEnabled")
-                .and_then(|v| v.as_bool())
-                .map(|b| if b { "enabled" } else { "disabled" }.to_string()),
-            ip_discovery: props
-                .get("IpDiscovery")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            network_type: props
-                .get("NetworkType")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            transit_encryption_mode: props
-                .get("TransitEncryptionMode")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            num_node_groups,
-            configuration_endpoint_address: configuration_endpoint.clone(),
-            configuration_endpoint_port: configuration_endpoint.as_ref().map(|_| port),
-            replicas_per_node_group,
-            auth_token: props
-                .get("AuthToken")
-                .and_then(|v| v.as_str())
-                .filter(|s| !s.is_empty())
-                .map(String::from),
-            port,
-            notification_topic_arn: props
-                .get("NotificationTopicArn")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            cluster_mode: props
-                .get("ClusterMode")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            data_tiering_enabled: props.get("DataTieringEnabled").and_then(|v| v.as_bool()),
-            notification_topic_status: None,
-            cache_parameter_group_name: props
-                .get("CacheParameterGroupName")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            cache_subnet_group_name: props
-                .get("CacheSubnetGroupName")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            security_group_ids: props
-                .get("SecurityGroupIds")
-                .and_then(|v| v.as_array())
-                .map(|arr| {
-                    arr.iter()
-                        .filter_map(|v| v.as_str().map(String::from))
-                        .collect()
-                })
-                .unwrap_or_default(),
-            preferred_maintenance_window: props
-                .get("PreferredMaintenanceWindow")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            snapshot_name: props
-                .get("SnapshotName")
-                .and_then(|v| v.as_str())
-                .map(String::from),
-            snapshot_arns: props
-                .get("SnapshotArns")
-                .and_then(|v| v.as_array())
-                .map(|arr| {
-                    arr.iter()
-                        .filter_map(|v| v.as_str().map(String::from))
-                        .collect()
-                })
-                .unwrap_or_default(),
-            auto_minor_version_upgrade: props
-                .get("AutoMinorVersionUpgrade")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(true),
-        };
-        state.replication_groups.insert(id.clone(), group);
-        drop(accounts);
-
-        if back_with_container {
-            self.pending_container_spawns.lock().push(
-                ContainerSpawnIntent::ElastiCacheReplicationGroup {
-                    replication_group_id: id.clone(),
-                },
-            );
-        }
-
-        let mut result = ProvisionResult::new(id.clone())
-            .with("Arn", arn)
-            .with("PrimaryEndPoint.Address", endpoint_address.clone())
-            .with("PrimaryEndPoint.Port", port.to_string())
-            .with("ReadEndPoint.Addresses", endpoint_address.clone())
-            .with("ReadEndPoint.Ports", port.to_string());
-        if let Some(cfg) = configuration_endpoint {
-            result = result
-                .with("ConfigurationEndPoint.Address", cfg)
-                .with("ConfigurationEndPoint.Port", port.to_string());
-        }
-        Ok(result)
-    }
-
-    fn delete_ec_replication_group(&self, physical_id: &str) -> Result<(), String> {
-        {
-            let mut accounts = self.elasticache_state.write();
-            let state = accounts.get_or_create(&self.account_id);
-            state.replication_groups.remove(physical_id);
-        }
-        if self.elasticache_runtime.is_some() {
-            self.pending_container_teardowns.lock().push(
-                ContainerTeardownIntent::ElastiCacheReplicationGroup {
-                    replication_group_id: physical_id.to_string(),
-                },
-            );
-        }
-        Ok(())
-    }
-
-    // --- CloudFront ---
-    //
-    // CloudFront is a global service that stores data under the default
-    // account bucket; mirror that here so SDK reads land on the same data
-    // CFN wrote.
-
-    fn create_cf_origin_access_identity(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("CloudFrontOriginAccessIdentityConfig")
-            .ok_or("CloudFrontOriginAccessIdentityConfig is required")?;
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let caller_reference = format!("cfn-{}", resource.logical_id);
-
-        let id = format!("E{}", fakecloud_core::ids::short_id(13).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-        let s3_canonical_user_id = format!(
-            "{:0<64}",
-            Uuid::new_v4().simple().to_string().to_lowercase()
-        );
-
-        let oai = StoredOriginAccessIdentity {
-            id: id.clone(),
-            etag,
-            s3_canonical_user_id: s3_canonical_user_id.clone(),
-            config: CloudFrontOriginAccessIdentityConfig {
-                caller_reference,
-                comment,
-            },
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.origin_access_identities.insert(id.clone(), oai);
-
-        Ok(ProvisionResult::new(id.clone())
-            .with("Id", id)
-            .with("S3CanonicalUserId", s3_canonical_user_id))
-    }
-
-    fn delete_cf_origin_access_identity(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.origin_access_identities.remove(physical_id);
-        Ok(())
-    }
-
-    /// Provision an `AWS::CloudFront::Distribution`. Reads
-    /// DistributionConfig.Origins/DefaultCacheBehavior/etc. and persists
-    /// a StoredDistribution in CloudFront state. CFN's Origins property
-    /// is a flat array, so we wrap it back into the wire shape with a
-    /// quantity + Items.Origin nesting.
-    fn create_cf_distribution(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let cfg = resource
-            .properties
-            .get("DistributionConfig")
-            .ok_or_else(|| "DistributionConfig is required".to_string())?;
-
-        // CFN Origins is a flat JSON array; the wire shape is
-        // { Quantity, Items: { Origin: [...] } }. Translate. CustomOriginConfig
-        // uses AWS's HTTPPort/HTTPSPort casing, which the model now accepts
-        // natively (see CustomOriginConfig in fakecloud-cloudfront::model), so no
-        // field patching is needed here.
-        let origin_entries: Vec<Origin> = cfg
-            .get("Origins")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| "DistributionConfig.Origins is required".to_string())?
-            .iter()
-            .map(|o| {
-                serde_json::from_value::<Origin>(o.clone())
-                    .map_err(|e| format!("Invalid Origin entry: {e}"))
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        if origin_entries.is_empty() {
-            return Err("DistributionConfig.Origins must contain at least one origin".to_string());
-        }
-        let origins = Origins {
-            quantity: origin_entries.len() as i32,
-            items: Some(OriginItems {
-                origin: origin_entries,
-            }),
-        };
-
-        let dcb_value = cfg
-            .get("DefaultCacheBehavior")
-            .ok_or_else(|| "DistributionConfig.DefaultCacheBehavior is required".to_string())?;
-        let default_cache_behavior: DefaultCacheBehavior =
-            serde_json::from_value(dcb_value.clone())
-                .map_err(|e| format!("Invalid DefaultCacheBehavior: {e}"))?;
-
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let enabled = cfg.get("Enabled").and_then(|v| v.as_bool()).unwrap_or(true);
-        let price_class = cfg
-            .get("PriceClass")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let http_version = cfg
-            .get("HttpVersion")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let is_ipv6_enabled = cfg.get("IPV6Enabled").and_then(|v| v.as_bool());
-        let default_root_object = cfg
-            .get("DefaultRootObject")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-        let web_acl_id = cfg
-            .get("WebACLId")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        let viewer_certificate: Option<ViewerCertificate> = cfg
-            .get("ViewerCertificate")
-            .map(|v| serde_json::from_value(v.clone()))
-            .transpose()
-            .map_err(|e| format!("Invalid ViewerCertificate: {e}"))?;
-
-        let caller_reference = format!("cfn-{}-{}", resource.logical_id, Uuid::new_v4().simple());
-
-        let mut config = DistributionConfig {
-            caller_reference,
-            comment,
-            enabled,
-            origins,
-            default_cache_behavior,
-            ..Default::default()
-        };
-        config.price_class = price_class;
-        config.http_version = http_version;
-        config.is_ipv6_enabled = is_ipv6_enabled;
-        config.default_root_object = default_root_object;
-        config.web_acl_id = web_acl_id;
-        config.viewer_certificate = viewer_certificate;
-
-        // Mint distribution id + ARN + domain in the same shape the
-        // CloudFront service uses.
-        let id_suffix: String = Uuid::new_v4()
-            .simple()
-            .to_string()
-            .chars()
-            .take(13)
-            .collect::<String>()
-            .to_uppercase();
-        let id = format!("E{id_suffix}");
-        let etag_suffix: String = Uuid::new_v4()
-            .simple()
-            .to_string()
-            .chars()
-            .take(7)
-            .collect::<String>()
-            .to_uppercase();
-        let etag = format!("E{etag_suffix}");
-        let domain_name = format!("{}.cloudfront.net", id.to_lowercase());
-        let arn = format!(
-            "arn:aws:cloudfront::{}:distribution/{}",
-            self.account_id, id
-        );
-
-        let stored = StoredDistribution {
-            id: id.clone(),
-            arn: arn.clone(),
-            // CloudFront flips this to Deployed on the first GetDistribution
-            // poll, matching the rest of the service.
-            status: "InProgress".to_string(),
-            last_modified_time: Utc::now(),
-            domain_name: domain_name.clone(),
-            in_progress_invalidation_batches: 0,
-            etag,
-            config,
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.distributions.insert(id.clone(), stored);
-        Ok(ProvisionResult::new(id.clone())
-            .with("Id", id)
-            .with("DomainName", domain_name)
-            .with("Arn", arn))
-    }
-
-    fn delete_cf_distribution(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.distributions.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_origin_access_control(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("OriginAccessControlConfig")
-            .ok_or("OriginAccessControlConfig is required")?;
-        let name = cfg
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("OriginAccessControlConfig.Name is required")?
-            .to_string();
-        let signing_protocol = cfg
-            .get("SigningProtocol")
-            .and_then(|v| v.as_str())
-            .unwrap_or("sigv4")
-            .to_string();
-        let signing_behavior = cfg
-            .get("SigningBehavior")
-            .and_then(|v| v.as_str())
-            .unwrap_or("always")
-            .to_string();
-        let origin_type = cfg
-            .get("OriginAccessControlOriginType")
-            .and_then(|v| v.as_str())
-            .ok_or("OriginAccessControlConfig.OriginAccessControlOriginType is required")?
-            .to_string();
-        let description = cfg
-            .get("Description")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        let id = format!("E{}", fakecloud_core::ids::short_id(13).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-        let oac = StoredOriginAccessControl {
-            id: id.clone(),
-            etag,
-            config: OriginAccessControlConfig {
-                name,
-                description,
-                signing_protocol,
-                signing_behavior,
-                origin_access_control_origin_type: origin_type,
-            },
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.origin_access_controls.insert(id.clone(), oac);
-
-        Ok(ProvisionResult::new(id.clone()).with("Id", id))
-    }
-
-    fn delete_cf_origin_access_control(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.origin_access_controls.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_public_key(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("PublicKeyConfig")
-            .ok_or("PublicKeyConfig is required")?;
-        let name = cfg
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("PublicKeyConfig.Name is required")?
-            .to_string();
-        let encoded_key = cfg
-            .get("EncodedKey")
-            .and_then(|v| v.as_str())
-            .ok_or("PublicKeyConfig.EncodedKey is required")?
-            .to_string();
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-        let caller_reference = cfg
-            .get("CallerReference")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let caller_reference = if caller_reference.is_empty() {
-            format!("cfn-{}", resource.logical_id)
-        } else {
-            caller_reference
-        };
-
-        let id = format!("K{}", fakecloud_core::ids::short_id(13).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-
-        let pk = StoredPublicKey {
-            id: id.clone(),
-            etag,
-            created_time: Utc::now(),
-            config: PublicKeyConfig {
-                caller_reference,
-                name,
-                encoded_key,
-                comment,
-            },
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.public_keys.insert(id.clone(), pk);
-
-        Ok(ProvisionResult::new(id.clone()).with("Id", id))
-    }
-
-    fn delete_cf_public_key(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.public_keys.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_key_group(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("KeyGroupConfig")
-            .ok_or("KeyGroupConfig is required")?;
-        let name = cfg
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("KeyGroupConfig.Name is required")?
-            .to_string();
-        let items: Vec<String> = cfg
-            .get("Items")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        let id = format!("KG{}", fakecloud_core::ids::short_id(12).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-
-        let kg = StoredKeyGroup {
-            id: id.clone(),
-            etag,
-            last_modified_time: Utc::now(),
-            config: KeyGroupConfig {
-                name,
-                items: KeyGroupItems { public_key: items },
-                comment,
-            },
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.key_groups.insert(id.clone(), kg);
-
-        Ok(ProvisionResult::new(id.clone()).with("Id", id))
-    }
-
-    fn delete_cf_key_group(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.key_groups.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_function(&self, resource: &ResourceDefinition) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let name = props
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("Name is required")?
-            .to_string();
-        let function_code = props
-            .get("FunctionCode")
-            .and_then(|v| v.as_str())
-            .ok_or("FunctionCode is required")?
-            .to_string();
-        let cfg = props
-            .get("FunctionConfig")
-            .ok_or("FunctionConfig is required")?;
-        let runtime = cfg
-            .get("Runtime")
-            .and_then(|v| v.as_str())
-            .unwrap_or("cloudfront-js-2.0")
-            .to_string();
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        let id = format!("FN{}", fakecloud_core::ids::short_id(12).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-        let function_arn =
-            Arn::global("cloudfront", &self.account_id, &format!("function/{name}")).to_string();
-
-        let now = Utc::now();
-        let func = StoredFunction {
-            name: name.clone(),
-            etag,
-            status: "UNPUBLISHED".to_string(),
-            stage: "DEVELOPMENT".to_string(),
-            function_arn: function_arn.clone(),
-            created_time: now,
-            last_modified_time: now,
-            config: FunctionConfig {
-                comment,
-                runtime,
-                key_value_store_associations: None,
-            },
-            function_code,
-            live_function_code: None,
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        // Use the function's ARN/name as the registry key so subsequent
-        // operations (Get/Update/Delete) keyed by name resolve.
-        state.functions.insert(name.clone(), func);
-
-        Ok(ProvisionResult::new(name.clone())
-            .with("FunctionARN", function_arn)
-            .with("FunctionMetadata.FunctionARN", id)
-            .with("Stage", "DEVELOPMENT"))
-    }
-
-    fn delete_cf_function(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.functions.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_cache_policy(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("CachePolicyConfig")
-            .ok_or("CachePolicyConfig is required")?;
-        let name = cfg
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("CachePolicyConfig.Name is required")?
-            .to_string();
-        let min_ttl = cfg
-            .get("MinTTL")
-            .and_then(|v| {
-                v.as_i64()
-                    .or_else(|| v.as_str().and_then(|s| s.parse::<i64>().ok()))
-            })
-            .unwrap_or(0);
-        let default_ttl = cfg.get("DefaultTTL").and_then(|v| {
-            v.as_i64()
-                .or_else(|| v.as_str().and_then(|s| s.parse::<i64>().ok()))
-        });
-        let max_ttl = cfg.get("MaxTTL").and_then(|v| {
-            v.as_i64()
-                .or_else(|| v.as_str().and_then(|s| s.parse::<i64>().ok()))
-        });
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        let id = format!("CP{}", fakecloud_core::ids::short_id(12).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-
-        let cache_policy = StoredCachePolicy {
-            id: id.clone(),
-            etag,
-            last_modified_time: Utc::now(),
-            config: CachePolicyConfig {
-                comment,
-                name,
-                default_ttl,
-                max_ttl,
-                min_ttl,
-                parameters_in_cache_key_and_forwarded_to_origin: None,
-            },
-            policy_type: "custom".to_string(),
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.cache_policies.insert(id.clone(), cache_policy);
-
-        Ok(ProvisionResult::new(id.clone()).with("Id", id))
-    }
-
-    fn delete_cf_cache_policy(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.cache_policies.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_origin_request_policy(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("OriginRequestPolicyConfig")
-            .ok_or("OriginRequestPolicyConfig is required")?;
-        let name = cfg
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("OriginRequestPolicyConfig.Name is required")?
-            .to_string();
-        let header_behavior = cfg
-            .get("HeadersConfig")
-            .and_then(|v| v.get("HeaderBehavior"))
-            .and_then(|v| v.as_str())
-            .unwrap_or("none")
-            .to_string();
-        let cookie_behavior = cfg
-            .get("CookiesConfig")
-            .and_then(|v| v.get("CookieBehavior"))
-            .and_then(|v| v.as_str())
-            .unwrap_or("none")
-            .to_string();
-        let query_string_behavior = cfg
-            .get("QueryStringsConfig")
-            .and_then(|v| v.get("QueryStringBehavior"))
-            .and_then(|v| v.as_str())
-            .unwrap_or("none")
-            .to_string();
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        let id = format!("ORP{}", fakecloud_core::ids::short_id(11).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-
-        let policy = StoredOriginRequestPolicy {
-            id: id.clone(),
-            etag,
-            last_modified_time: Utc::now(),
-            config: OriginRequestPolicyConfig {
-                comment,
-                name,
-                headers_config: OriginRequestPolicyHeadersConfig {
-                    header_behavior,
-                    headers: None,
-                },
-                cookies_config: OriginRequestPolicyCookiesConfig {
-                    cookie_behavior,
-                    cookies: None,
-                },
-                query_strings_config: OriginRequestPolicyQueryStringsConfig {
-                    query_string_behavior,
-                    query_strings: None,
-                },
-            },
-            policy_type: "custom".to_string(),
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.origin_request_policies.insert(id.clone(), policy);
-
-        Ok(ProvisionResult::new(id.clone()).with("Id", id))
-    }
-
-    fn delete_cf_origin_request_policy(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.origin_request_policies.remove(physical_id);
-        Ok(())
-    }
-
-    fn create_cf_response_headers_policy(
-        &self,
-        resource: &ResourceDefinition,
-    ) -> Result<ProvisionResult, String> {
-        let props = &resource.properties;
-        let cfg = props
-            .get("ResponseHeadersPolicyConfig")
-            .ok_or("ResponseHeadersPolicyConfig is required")?;
-        let name = cfg
-            .get("Name")
-            .and_then(|v| v.as_str())
-            .ok_or("ResponseHeadersPolicyConfig.Name is required")?
-            .to_string();
-        let comment = cfg
-            .get("Comment")
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        let id = format!("RHP{}", fakecloud_core::ids::short_id(11).to_uppercase());
-        let etag = format!("E{}", fakecloud_core::ids::short_id(7).to_uppercase());
-
-        let policy = StoredResponseHeadersPolicy {
-            id: id.clone(),
-            etag,
-            last_modified_time: Utc::now(),
-            config: ResponseHeadersPolicyConfig {
-                comment,
-                name,
-                cors_config: None,
-                security_headers_config: None,
-                server_timing_headers_config: None,
-                custom_headers_config: None,
-                remove_headers_config: None,
-            },
-            policy_type: "custom".to_string(),
-        };
-
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.response_headers_policies.insert(id.clone(), policy);
-
-        Ok(ProvisionResult::new(id.clone()).with("Id", id))
-    }
-
-    fn delete_cf_response_headers_policy(&self, physical_id: &str) -> Result<(), String> {
-        let mut accounts = self.cloudfront_state.write();
-        let state = accounts.entry("000000000000");
-        state.response_headers_policies.remove(physical_id);
-        Ok(())
-    }
 
     fn parse_athena_tags(value: Option<&serde_json::Value>) -> BTreeMap<String, String> {
         let mut out = BTreeMap::new();
